fix casting

+++ server/src/main/java/io/druid/segment/realtime/plumber/FlushingPlumber.java
@@ -96,7 +96,7 @@ public class FlushingPlumber extends RealtimePlumber
     log.info(
         "Abandoning segment %s at %s",
         sink.getSegment().getIdentifier(),
-        new DateTime().plusMillis(flushDuration.getMillis())
+        new DateTime().plusMillis((int) flushDuration.getMillis())
     );
 
     ScheduledExecutors.scheduleWithFixedDelay(
